requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
xrpl-py==4.0.0
pydantic-settings==2.6.1
python-dotenv==1.0.1
structlog==24.4.0
slowapi==0.1.9
redis==5.2.1
httpx==0.28.1
